getaddrinfo should be done in resolv
